requirements.txt
asgiref==3.5.2
Brotli==1.0.9
cffi==1.15.1
click==8.1.3
cssselect2==0.7.0
Django==4.1.3
django-chartjs==2.3.0
django-stdimage==6.0.1
django-weasyprint==2.2.0
fonttools==4.38.0
html5lib==1.1
joblib==1.2.0
nltk==3.7
Pillow==9.3.0
pycparser==2.21
pydyf==0.5.0
pyphen==0.13.1
regex==2022.10.31
six==1.16.0
slugify==0.0.1
sqlparse==0.4.3
textblob==0.17.1
tinycss2==1.2.1
tqdm==4.64.1
weasyprint==57.1
webencodings==0.5.1
zopfli==0.2.2
